Converted Admin Page Finder from Perl to Python3 ,
he main script is an admin page finder tool written in Python. It allows the user to search for admin login pages on a target domain by performing either a Light Scan or a Deep Scan.

Upon running the script, the user is prompted to enter the target domain. The script then normalizes and validates the URL. After that, the user is presented with a menu to choose between a Light Scan or a Deep Scan.

During the scan, the script retrieves a list of paths to check based on the chosen scan type. For a Light Scan, the paths are obtained from a file named PATH.py, while for a Deep Scan, the paths are obtained from a file named DEEPPATH.py.

The script then initiates the scan and displays relevant information, including the target domain, the total number of paths to check, and the progress of the scan. Each path is checked by combining it with the target domain and making an HTTP request to validate its existence.

If an admin login page is found, its URL is displayed to the user. At the end of the scan, the script provides a summary of the results, including the number of admin login pages found.

The main script has been enhanced with visual improvements, such as colorful ASCII art banners at startup and formatting adjustments to highlight important information. The updates aim to make the script more visually appealing and provide a modern look and feel while delivering comprehensive information to the user during runtime.